refactor folders architecture in front end components
