+++ front/src/components/authentication/SignInPage.tsx
@@ -2,9 +2,9 @@ import React, { useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { ApolloError, useMutation } from '@apollo/client';
 import styled from 'styled-components';
-import Context from './context/Context';
-import ErrorMessage from './ErrorMessage';
-import { LOGIN, ME } from '../utils/apollo';
+import Context from '../context/Context';
+import ErrorMessage from '../core/ErrorMessage';
+import { LOGIN, ME } from '../../utils/apollo';
 
 const Wrapper = styled.div`
 	display: flex;
