+++ front/src/components/backOfffice/FormMasterbackOffice.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import ErrorMessage from '../ErrorMessage';
+import ErrorMessage from '../core/ErrorMessage';
 
 const Container = styled.div`
 	display: 'flex';
