+++ front/src/components/templates/Footer.styled.tsx
