+++ front/src/components/templates/Layout.styled.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import Header from './Header.styled';
-import Footer from './Footer.styled';
+// import Footer from './templates/Footer.styled';
 
 const Container = styled.div``;
 const App = styled.div`
