+++ front/src/components/timeline/Timeline.styled.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { useQuery } from '@apollo/client';
 import { Chrono } from 'react-chrono';
 import styled from 'styled-components';
-import convertDate from '../utils/convertDate';
-import ErrorMessage from './ErrorMessage';
-import { GET_COURSES } from '../utils/apollo';
+import convertDate from '../../utils/convertDate';
+import ErrorMessage from '../core/ErrorMessage';
+import { GET_COURSES } from '../../utils/apollo';
 
 const TimelineContent = styled.div`
 	width: 100%;
