+++ front/src/tests/Timeline.test.tsx
@@ -2,12 +2,13 @@ import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 import { waitFor, screen, render } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
-import { GET_COURSES_QUERY, Timeline } from '../components/Timeline.styled';
+import { Timeline } from '../components/timeline/Timeline.styled';
+import { GET_COURSES } from '../utils/apollo';
 
 const mocks = [
 	{
 		request: {
-			query: GET_COURSES_QUERY,
+			query: GET_COURSES,
 		},
 		result: {
 			data: {
